refactor: huge refactoring not working now
