+++ jina/parsers/helper.py
@@ -290,10 +290,6 @@ def _set_gateway_uses(args: 'argparse.Namespace'):
 
 
 def _update_gateway_args(args: 'argparse.Namespace'):
-    from jina.helper import random_ports
-
-    if not args.port:
-        args.port = random_ports(len(args.protocol))
     _set_gateway_uses(args)
