+++ jina/serve/runtimes/gateway/composite/__init__.py
@@ -1,3 +1,8 @@
-from jina.serve.runtimes.gateway.composite.gateway import CompositeGateway
+from jina.serve.runtimes.gateway.gateway import BaseGateway
+from jina.serve.runtimes.servers.composite import CompositeServer
 
 __all__ = ['CompositeGateway']
+
+
+class CompositeGateway(CompositeServer, BaseGateway):
+    pass
