+++ jina/serve/runtimes/gateway/gateway.py
@@ -1,0 +1,50 @@
+from jina.jaml import JAMLCompatible
+from jina.serve.helper import store_init_kwargs, wrap_func
+
+__all__ = ['BaseGateway']
+
+
+class GatewayType(type(JAMLCompatible), type):
+    """The class of Gateway type, which is the metaclass of :class:`BaseGateway`."""
+
+    def __new__(cls, *args, **kwargs):
+        """
+        # noqa: DAR101
+        # noqa: DAR102
+
+        :return: Gateway class
+        """
+        _cls = super().__new__(cls, *args, **kwargs)
+        return cls.register_class(_cls)
+
+    @staticmethod
+    def register_class(cls):
+        """
+        Register a class.
+
+        :param cls: The class.
+        :return: The class, after being registered.
+        """
+
+        reg_cls_set = getattr(cls, '_registered_class', set())
+
+        cls_id = f'{cls.__module__}.{cls.__name__}'
+        if cls_id not in reg_cls_set:
+            reg_cls_set.add(cls_id)
+            setattr(cls, '_registered_class', reg_cls_set)
+            wrap_func(
+                cls,
+                ['__init__'],
+                store_init_kwargs,
+                taboo={'self', 'args', 'kwargs', 'runtime_args'},
+            )
+        return cls
+
+
+class BaseGateway(JAMLCompatible, metaclass=GatewayType):
+    """
+    The base class of all custom Gateways, can be used to build a custom interface to a Jina Flow that supports
+    gateway logic
+
+    :class:`jina.Gateway` as an alias for this class.
+    """
