+++ jina/serve/runtimes/gateway/grpc/__init__.py
@@ -1,3 +1,9 @@
-from jina.serve.runtimes.gateway.grpc.gateway import GRPCGateway
+from jina.serve.runtimes.gateway.gateway import BaseGateway
+from jina.serve.runtimes.servers.grpc import GRPCServer
 
 __all__ = ['GRPCGateway']
+
+
+class GRPCGateway(GRPCServer, BaseGateway):
+    pass
+
