+++ jina/serve/runtimes/gateway/http/__init__.py
@@ -1,3 +1,14 @@
-from jina.serve.runtimes.gateway.http.gateway import HTTPGateway
+from jina.serve.runtimes.gateway.gateway import BaseGateway
+from jina.serve.runtimes.servers.http import FastAPIBaseServer
 
-__all__ = ['HTTPGateway']
+__all__ = ['HTTPGateway', 'FastAPIBaseGateway']
+
+
+class FastAPIBaseGateway(FastAPIBaseServer, BaseGateway):
+    pass
+
+
+class HTTPGateway(FastAPIBaseGateway):
+    @property
+    def app(self):
+        return self._request_handler._http_fastapi_default_app()
