+++ jina/serve/runtimes/gateway/websocket_fastapi_app.py
@@ -23,10 +23,10 @@ def _fits_ws_close_msg(msg: str):
 
 
 def get_fastapi_app(
-    streamer: 'GatewayStreamer',
-    logger: 'JinaLogger',
-    tracing: Optional[bool] = None,
-    tracer_provider: Optional['trace.TracerProvider'] = None,
+        streamer: 'GatewayStreamer',
+        logger: 'JinaLogger',
+        tracing: Optional[bool] = None,
+        tracer_provider: Optional['trace.TracerProvider'] = None,
 ):
     """
     Get the app from FastAPI as the Websocket interface.
@@ -102,7 +102,7 @@ def get_fastapi_app(
                 pass
 
         async def send(
-            self, websocket: WebSocket, data: Union[DataRequest, StatusMessage]
+                self, websocket: WebSocket, data: Union[DataRequest, StatusMessage]
         ) -> None:
             subprotocol = self.protocol_dict[self.get_client(websocket)]
             if subprotocol == WebsocketSubProtocols.JSON:
@@ -156,7 +156,7 @@ def get_fastapi_app(
 
     @app.websocket('/')
     async def websocket_endpoint(
-        websocket: WebSocket, response: Response
+            websocket: WebSocket, response: Response
     ):  # 'response' is a FastAPI response, not a Jina response
         await manager.connect(websocket)
 
@@ -226,7 +226,7 @@ def get_fastapi_app(
     @app.get(
         path='/dry_run',
         summary='Get the readiness of Jina Flow service, sends an empty DocumentArray to the complete Flow to '
-        'validate connectivity',
+                'validate connectivity',
         response_model=PROTO_TO_PYDANTIC_MODELS.StatusProto,
     )
     async def _dry_run_http():
@@ -256,7 +256,7 @@ def get_fastapi_app(
 
     @app.websocket('/dry_run')
     async def websocket_endpoint(
-        websocket: WebSocket, response: Response
+            websocket: WebSocket, response: Response
     ):  # 'response' is a FastAPI response, not a Jina response
         from jina.proto import jina_pb2
         from jina.serve.executors import __dry_run_endpoint__
@@ -266,11 +266,11 @@ def get_fastapi_app(
         da = DocumentArray([])
         try:
             async for _ in streamer.rpc_stream(
-                request_iterator=request_generator(
-                    exec_endpoint=__dry_run_endpoint__,
-                    data=da,
-                    data_type=DataInputType.DOCUMENT,
-                )
+                    request_iterator=request_generator(
+                        exec_endpoint=__dry_run_endpoint__,
+                        data=da,
+                        data_type=DataInputType.DOCUMENT,
+                    )
             ):
                 pass
             status_message = StatusMessage()
