+++ jina/serve/runtimes/worker/request_handling.py
@@ -458,6 +458,7 @@ class WorkerRequestHandler:
     def _set_result(self, requests, return_data, docs):
         # assigning result back to request
         if return_data is not None:
+            print(f' return_data {return_data}')
             if isinstance(return_data, DocumentArray):
                 docs = return_data
             elif isinstance(return_data, dict):
